+++ src/gcp/cloudsql/permissionsSetup.ts
@@ -16,11 +16,13 @@ import { iamUserIsCSQLAdmin } from "./cloudsqladmin";
 import { logger } from "../../logger";
 import { confirm } from "../../prompt";
 import { FirebaseError } from "../../error";
-import { needProjectNumber } from "../../projectUtils";
+import { needProjectId, needProjectNumber } from "../../projectUtils";
 import { executeSqlCmdsAsIamUser, executeSqlCmdsAsSuperUser, getIAMUser } from "./connect";
 import { concat } from "lodash";
 import { getDataConnectP4SA, toDatabaseUser } from "./connect";
 import * as utils from "../../utils";
+// TODO: Update the import path below to the correct location of cloudsqladmin if it exists elsewhere.
+import * as cloudSqlAdminClient from "./cloudsqladmin";
 
 export type TableMetadata = {
   name: string;
@@ -445,5 +447,27 @@ export async function brownfieldSqlSetup(
     brownfieldSetupCmds,
     silent,
     /** transaction=*/ true,
+  );
+}
+
+export async function grantRoleTo(
+  options: Options,
+  instanceId: string,
+  databaseId: string,
+  role: string,
+  email: string,
+): Promise<void> {
+  // Upsert new user account into the database.
+  const projectId = needProjectId(options);
+  const { user, mode } = toDatabaseUser(email);
+  await cloudSqlAdminClient.createUser(projectId, instanceId, mode, user);
+
+  const fdcSqlRole = fdcSqlRoleMap[role as keyof typeof fdcSqlRoleMap](databaseId);
+  await executeSqlCmdsAsSuperUser(
+    options,
+    instanceId,
+    databaseId,
+    /** cmds= */ [`GRANT "${fdcSqlRole}" TO "${user}"`],
+    /** silent= */ false,
   );
 }
